stage2/requirements.txt
numpy
boto3
opencv-python
facenet-pytorch
pillow